+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Please see https://github.com/elelabdev/net-commander/releases for the latest re
 All notable changes to the "net-commander" extension will be documented in this file.
 
 <br><br>
+## 2025-06-20 - [0.0.3]
+### Added
+- Minor fixes for Root Cause Analysis for Microsoft Azure
 
 ## 2025-05-16 - [0.0.2]
 ### Added